requirements.txt
numpy==2.3.2
onnxruntime==1.22.0
openai==1.97.1
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==9.5.0
Pillow==11.3.0
protobuf==6.31.1
PySide6==6.9.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PyYAML==6.0.2
PyYAML==6.0.2
simple_lama_inpainting==0.1.2
ultralytics==8.3.109
ultralytics==8.3.158
